+++ src/ray.rs
@@ -2,6 +2,6 @@ use nalgebra::{Point3, Vector3};
 
 #[derive(Debug)]
 pub struct Ray {
-    pub start: Point3<f32>,
-    pub direction: Vector3<f32>,
+    pub start: Point3<f64>,
+    pub direction: Vector3<f64>,
 }
